+++ catscripts/Entertainment/Video_games/MMORPGs/Eve_Online/EveOnline.sh
@@ -2,15 +2,23 @@
 
 export KEYWORDS_MMORPG_EVE="Eve(| )Online"
 
-egrep -i "$KEYWORDS_MMORPG_EVE" newpages.txt >> EveOnline.txt
-
-EVE=$(stat --print=%s EveOnline.txt)
-
-if [ $EVE -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="EveOnline.txt"
-  export CATNAME="Eve Online"
-  $CATEGORIZE
-fi
 
-rm EveOnline.txt
+  debug_start "Eve Online"
+
+  EVE=$(egrep -i "$KEYWORDS_MMORPG_EVE" newpages.txt)
+
+  if [ "$EVE" != "" ];
+  then
+    printf "%s" "$EVE" > EveOnline.txt
+    export CATFILE="EveOnline.txt"
+    export CATNAME="Eve Online"
+    $CATEGORIZE
+    rm EveOnline.txt
+    unset EVE
+  fi
+
+  debug_end "Eve Online"
+
+fi
