+++ catscripts/Entertainment/Video_games/MMORPGs/Karos_Online/KarosOnline.sh
@@ -2,15 +2,23 @@
 
 export KEYWORDS_MMORPG_KAROS="\bKaros\b"
 
-egrep -i "$KEYWORDS_MMORPG_KAROS" newpages.txt >> KarosOnline.txt
-
-KAROS=$(stat --print=%s KarosOnline.txt)
-
-if [ $KAROS -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="KarosOnline.txt"
-  export CATNAME="Karos Online"
-  $CATEGORIZE
-fi
 
-rm KarosOnline.txt
+  debug_start "Karos Online"
+
+  KAROS=$(egrep -i "$KEYWORDS_MMORPG_KAROS" newpages.txt)
+
+  if [ "$KAROS" != "" ];
+  then
+    printf "%s" "$KAROS" > KarosOnline.txt
+    export CATFILE="KarosOnline.txt"
+    export CATNAME="Karos Online"
+    $CATEGORIZE
+    rm KarosOnline.txt
+    unset KAROS
+  fi
+
+  debug_end "Karos Online"
+
+fi
