+++ catscripts/Entertainment/Video_games/Video_game_consoles/Microsoft/XboxLive.sh
@@ -2,15 +2,23 @@
 
 KEYWORDS_XBOXLIVE="X(|-)box(| )Live"
 
-egrep -i "$KEYWORDS_XBOXLIVE" newpages.txt >> XboxLive.txt
-
-XBOXLIVE=$(stat --print=%s XboxLive.txt)
-
-if [ $XBOXLIVE -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="XboxLive.txt"
-  export CATNAME="Xbox Live"
-  $CATEGORIZE
-fi
+  
+  debug_start "Xbox Live"
 
-rm XboxLive.txt
+  XBOXLIVE=$(egrep -i "$KEYWORDS_XBOXLIVE" newpages.txt)
+
+  if [ "$XBOXLIVE" != "" ];
+  then
+    printf "%s" "$XBOXLIVE" > XboxLive.txt
+    export CATFILE="XboxLive.txt"
+    export CATNAME="Xbox Live"
+    $CATEGORIZE
+    rm XboxLive.txt
+    unset XBOXLIVE
+  fi
+
+  debug_end "Xbox Live"
+
+fi
