+++ catscripts/Entertainment/Video_games/Video_games_by_series/LittleBigPlanet_series/LittleBigPlanet.sh
@@ -1,14 +1,24 @@
 #!/bin/bash
 
-egrep -i 'LittleBigPlanet|Little Big Planet|\bLBP\b' newpages.txt >> LittleBigPlanet.txt
+KEYWORDS_LITTLEBIGPLANET="Little(| )Big(| )Planet|\bLBP\b"
 
-PLANET=$(stat --print=%s LittleBigPlanet.txt)
-
-if [ $PLANET -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="LittleBigPlanet.txt"
-  export CATNAME="LittleBigPlanet series"
-  $CATEGORIZE
-fi
 
-rm LittleBigPlanet.txt
+  debug_start "LittleBigPlanet series"
+
+  PLANET=$(egrep -i "$KEYWORDS_LITTLEBIGPLANET" newpages.txt)
+
+  if [ "$PLANET" != "" ];
+  then
+    printf "%s" "$PLANET" > LittleBigPlanet.txt
+    export CATFILE="LittleBigPlanet.txt"
+    export CATNAME="LittleBigPlanet series"
+    $CATEGORIZE
+    rm LittleBigPlanet.txt
+    unset PLANET
+  fi
+
+  debug_end "LittleBigPlanet series"
+
+fi
