+++ catscripts/Government/Countries/United_States/US_states/Mississippi/Mississippi.sh
@@ -2,16 +2,23 @@
 
 KEYWORDS_MISSISSIPPI="Mississippi"
 
-
-egrep -i "$KEYWORDS_MISSISSIPPI" newpages.txt >> Mississippi.txt
-
-MISSISSIPPI=$(stat --print=%s Mississippi.txt)
-
-if [ $MISSISSIPPI -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Mississippi.txt"
-  export CATNAME="Mississippi"
-  $CATEGORIZE
-fi
 
-rm Mississippi.txt
+  debug_start "Mississippi"
+
+  MISSISSIPPI=$(egrep -i "$KEYWORDS_MISSISSIPPI" newpages.txt)
+
+  if [ "$MISSISSIPPI" != "" ];
+  then
+    printf "%s" "$MISSISSIPPI" > Mississippi.txt
+    export CATFILE="Mississippi.txt"
+    export CATNAME="Mississippi"
+    $CATEGORIZE
+    rm Mississippi.txt
+    unset MISSISSIPPI
+  fi
+
+  debug_end "Mississippi"
+
+fi
