+++ catscripts/Government/Countries/United_States/US_states/Nebraska/Nebraska.sh
@@ -2,15 +2,23 @@
 
 KEYWORDS_NEBRASKA="Nebraska"
 
-egrep -i "$KEYWORDS_NEBRASKA" newpages.txt >> Nebraska.txt
-
-NEBRASKA=$(stat --print=%s Nebraska.txt)
-
-if [ $NEBRASKA -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Nebraska.txt"
-  export CATNAME="Nebraska"
-  $CATEGORIZE
-fi
 
-rm Nebraska.txt
+  debug_start "Nebraska"
+
+  NEBRASKA=$(egrep -i "$KEYWORDS_NEBRASKA" newpages.txt)
+
+  if [ "$NEBRASKA" != "" ];
+  then
+    printf "%s" "$NEBRASKA" > Nebraska.txt
+    export CATFILE="Nebraska.txt"
+    export CATNAME="Nebraska"
+    $CATEGORIZE
+    rm Nebraska.txt
+    unset NEBRASKA
+  fi
+
+  debug_end "Nebraska"
+
+fi
