+++ catscripts/Language/Non-English_languages/English_to_Navajo/EnglishtoNavajo.sh
@@ -1,14 +1,24 @@
 #!/bin/bash
 
-egrep -i '.+in Navajo$]|Navajo word for' newpages.txt >> EnglishtoNavajo.txt
+KEYWORDS_ENGLISHTONAVAJO=".+in Navajo$]|Navajo word for"
 
-ENGLISHTONAVAJO=$(stat --print=%s EnglishtoNavajo.txt)
-
-if [ $ENGLISHTONAVAJO -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="EnglishtoNavajo.txt"
-  export CATNAME="English to Navajo"
-  $CATEGORIZE
-fi
 
-rm EnglishtoNavajo.txt
+  debug_start "English to Navajo"
+
+  ENGLISHTONAVAJO=$(egrep -i "$KEYWORDS_ENGLISHTONAVAJO" newpages.txt)
+
+  if [ "$ENGLISHTONAVAJO" != "" ];
+  then
+    printf "%s" "$ENGLISHTONAVAJO" > EnglishtoNavajo.txt
+    export CATFILE="EnglishtoNavajo.txt"
+    export CATNAME="English to Navajo"
+    $CATEGORIZE
+    rm EnglishtoNavajo.txt
+    unset ENGLISHTONAVAJO
+  fi
+
+  debug_end "English to Navajo"
+
+fi
